Dev Snapshot

+++ trunk/hivewars/GameStateData.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 
 public class GameStateData implements Serializable{
     
-	short gameStateNum;        //clock time when this state occurred
+	int gameStateNum;        //clock time when this state occurred
 	ArrayList<Hive> hives; 	   //all hives and their statuses
 	ArrayList<Attack> attacks; //all attacks under progress
